synaps-idp-flask/requirements.txt
flask
pymysql
python-dotenv
pyjwt
